Add the footer bar with selected items
